+++ py/desisim/scripts/qa_zfind.py
@@ -9,11 +9,12 @@ Read fibermaps and zbest files to generate QA related to redshifts
 """
 
 import argparse
+from desisim.spec_qa import __qa_version__
 
 def parse(options=None):
 
 
-    parser = argparse.ArgumentParser(description="Generate QA on redshift for a production [v1.1]", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+    parser = argparse.ArgumentParser(description="Generate QA on redshift for a production [v{:s}]".format(__qa_version__), formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     parser.add_argument('--verbose', action = 'store_true',
         help = 'Provide verbose reporting of progress.')
     parser.add_argument('--load_simz_table', type = str, default = None, required=False,
@@ -22,9 +23,9 @@ def parse(options=None):
                         help = 'Override default path ($DESI_SPECTRO_REDUX/$SPECPROD) to processed data.')
     parser.add_argument('--rawdir', type = str, default = None, metavar = 'PATH',
                         help = 'Override default path ($DESI_SPECTRO_REDUX/$SPECPROD) to processed data.')
-    parser.add_argument('--qafile', type = str, default = None, required=False,
-                        help = 'path of QA file.')
-    parser.add_argument('--qafig_root', type=str, default=None, help = 'Root name (and path) of QA figure files')
+    parser.add_argument('--yaml_file', type = str, default = None, required=False,
+                        help = 'YAML file for debugging (primarily).')
+    parser.add_argument('--qafig_path', type=str, default=None, help = 'Path to where QA figure files are generated.  Default is specprod_dir+/QA')
     parser.add_argument('--write_simz_table', type=str, default=None, help = 'Write simz to this filename')
 
     if options is None:
@@ -51,6 +52,17 @@ def main(args):
     from desimodel.footprint import radec2pix
 
     log = get_logger()
+
+    # Initialize
+    if args.reduxdir is None:
+        specprod_dir = desispec.io.meta.specprod_root()
+    else:
+        specprod_dir = args.reduxdir
+
+    if args.qafig_path is not None:
+        qafig_path = args.qafig_path
+    else:
+        qafig_path = specprod_dir+'/QA/'
 
 
     if args.load_simz_table is not None:
@@ -138,25 +150,27 @@ def main(args):
     # Run stats
     log.info("Running stats..")
     summ_dict = dsqa_z.summ_stats(simz_tab)
-    if args.qafile is not None:
+    if args.yaml_file is not None:
         log.info("Generating yaml file: {:s}".format(args.qafile))
         # yamlify
         # Write yaml
-        with open(args.qafile, 'w') as outfile:
+        desispec.io.util.makepath(args.yaml_file)
+        with open(args.yaml_qafile, 'w') as outfile:
             outfile.write(yaml.dump(yamlify(meta), default_flow_style=False))
             outfile.write(yaml.dump(yamlify(summ_dict), default_flow_style=False))
 
-    if args.qafig_root is not None:
-        log.info("Generating QA files")
-        # Summary for dz of all types
-        outfile = args.qafig_root+'_dzsumm.png'
-        #dsqa_z.dz_summ(simz_tab, outfile=outfile)
-        # Summary of individual types
-        #outfile = args.qafig_root+'_summ_fig.png'
-        #dsqa_z.summ_fig(simz_tab, summ_dict, meta, outfile=outfile)
-        for objtype in ['BGS', 'MWS', 'ELG','LRG', 'QSO_T', 'QSO_L']:
-            outfile = args.qafig_root+'_zfind_{:s}.png'.format(objtype)
-            dsqa_z.obj_fig(simz_tab, objtype, summ_dict, outfile=outfile)
+    log.info("Generating QA files")
+    # Summary for dz of all types
+    outfile = qafig_path+'QA_dzsumm.png'
+    desispec.io.util.makepath(outfile)
+    dsqa_z.dz_summ(simz_tab, outfile=outfile)
+    # Summary of individual types
+    #outfile = args.qafig_root+'_summ_fig.png'
+    #dsqa_z.summ_fig(simz_tab, summ_dict, meta, outfile=outfile)
+    for objtype in ['BGS', 'MWS', 'ELG','LRG', 'QSO_T', 'QSO_L']:
+        outfile = qafig_path+'QA_zfind_{:s}.png'.format(objtype)
+        desispec.io.util.makepath(outfile)
+        dsqa_z.obj_fig(simz_tab, objtype, summ_dict, outfile=outfile)
 
 if __name__ == '__main__':
     main()