+++ py/desisim/spec_qa/__init__.py
@@ -5,4 +5,4 @@ desisim.spec_qa
 Tools for DESI QA with simulations
 It does not cover cosmology simulations.
 """
-__qa_version__ = '0.2.0'
+__qa_version__ = '0.2.1'