+++ doc/changes.rst
@@ -11,11 +11,15 @@ desisim change log
 * Fixes quickspectra (broken by desispec change) (`PR #306`_).
 * Fixes quickspectra random seed (never worked?) (`PR #306`_).
 * Improves pixsim_mpi performance (`PR #312`_).
+* Optionally do not wavelength resample simqso templates (`PR #310`_).
+* Default to basis templates v2.4 instead of 2.3
+* Minor edits to QA scripts and doc
 
 .. _`PR #302`: https://github.com/desihub/desisim/pull/302
 .. _`PR #303`: https://github.com/desihub/desisim/pull/303
 .. _`PR #306`: https://github.com/desihub/desisim/pull/306
 .. _`PR #312`: https://github.com/desihub/desisim/pull/312
+.. _`PR #310`: https://github.com/desihub/desisim/pull/310
 
 0.23.0 (2017-12-20)
 -------------------